Impl of product reservation with Redis listener

+++ common/src/main/java/pl/dminior8/common/dto/ProductDto.java
@@ -2,7 +2,6 @@ package pl.dminior8.common.dto;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -25,5 +24,5 @@ public class ProductDto {
     private int availableQuantity;
 
     @Min(0)
-    private float price;
+    private double price;
 }
